requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
annotated-types=0.7.0=pypi_0
anyio=4.4.0=pypi_0
asttokens=2.4.1=pypi_0
blas=1.0=mkl
bottleneck=1.3.7=py312he558020_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.7.2=haa95532_0
certifi=2024.6.2=pypi_0
cffi=1.16.0=py312h2bbff1b_1
charset-normalizer=3.3.2=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.2=pypi_0
debugpy=1.8.1=pypi_0
decorator=5.1.1=pypi_0
executing=2.0.1=pypi_0
expat=2.6.2=hd77b12b_0
filelock=3.13.1=py312haa95532_0
fsspec=2024.3.1=py312haa95532_0
grpcio=1.64.1=pypi_0
grpcio-tools=1.64.1=pypi_0
h11=0.14.0=pypi_0
h2=4.1.0=pypi_0
hpack=4.0.0=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.23.3=pypi_0
hyperframe=6.0.1=pypi_0
idna=3.7=pypi_0
intel-openmp=2023.1.0=h59b6b97_46320
ipykernel=6.29.4=pypi_0
ipython=8.25.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.4=py312haa95532_0
joblib=1.4.2=pypi_0
jupyter-client=8.6.2=pypi_0
jupyter-core=5.7.2=pypi_0
libffi=3.4.4=hd77b12b_1
libuv=1.44.2=h2bbff1b_0
lxml=5.2.2=pypi_0
markupsafe=2.1.3=py312h2bbff1b_0
matplotlib-inline=0.1.7=pypi_0
maturin=1.6.0=pypi_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py312h2bbff1b_1
mkl_fft=1.3.8=py312h2bbff1b_0
mkl_random=1.2.4=py312h59b6b97_0
mpmath=1.3.0=py312haa95532_0
nest-asyncio=1.6.0=pypi_0
networkx=3.2.1=py312haa95532_0
ninja=1.10.2=haa95532_5
ninja-base=1.10.2=h6d14046_5
numexpr=2.8.7=py312h96b7d27_0
numpy=1.26.4=py312hfd52020_0
numpy-base=1.26.4=py312h4dde369_0
openssl=3.0.14=h827c3e9_0
packaging=24.1=pypi_0
pandas=2.2.2=py312h0158946_0
parso=0.8.4=pypi_0
pika=1.3.2=pypi_0
pip=24.0=py312haa95532_0
platformdirs=4.2.2=pypi_0
portalocker=2.8.2=pypi_0
prompt-toolkit=3.0.47=pypi_0
protobuf=5.27.1=pypi_0
psutil=5.9.8=pypi_0
pure-eval=0.2.2=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=2.7.3=pypi_0
pydantic-core=2.18.4=pypi_0
pygments=2.18.0=pypi_0
pyodbc=5.0.1=py312hd77b12b_0
python=3.12.3=h1d929f7_1
python-crfsuite=0.9.10=pypi_0
python-dateutil=2.9.0post0=py312haa95532_2
python-docx=1.1.2=pypi_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytorch=2.2.0=cpu_py312h746a3fd_0
pytz=2024.1=py312haa95532_0
pyvi=0.1.1=pypi_0
pywin32=306=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=26.0.3=pypi_0
qdrant-client=1.9.1=pypi_0
regex=2024.5.15=pypi_0
requests=2.32.3=pypi_0
safetensors=0.4.3=pypi_0
scikit-learn=1.5.0=pypi_0
scipy=1.13.1=pypi_0
setuptools=69.5.1=py312haa95532_0
six=1.16.0=pyhd3eb1b0_1
sklearn-crfsuite=0.3.6=pypi_0
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h2bbff1b_0
stack-data=0.6.3=pypi_0
sympy=1.12=py312haa95532_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=h59b6b97_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h0416ee5_0
tokenizers=0.19.1=pypi_0
tornado=6.4.1=pypi_0
tqdm=4.66.4=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.41.2=pypi_0
typing-extensions=4.12.2=pypi_0
typing_extensions=4.11.0=py312haa95532_0
tzdata=2024a=h04d1e81_0
urllib3=2.2.1=pypi_0
vc=14.2=h2eaa2aa_1
vs2015_runtime=14.29.30133=h43f2093_3
wcwidth=0.2.13=pypi_0
wheel=0.43.0=py312haa95532_0
xz=5.4.6=h8cc25b3_1
zlib=1.2.13=h8cc25b3_1
